myarea now renders a new component and routes to /recipes
